refactor: birthday API를 get/post 단위로 통합
