+++ api/private/birthday/birthday.post.api.ts
@@ -9,3 +9,9 @@ export const postBirthdayLetter = async (
   const res = await api.post(`/birthday/letters/${receiverId}`, payload);
   return res.data;
 };
+
+// 🎉 생일 편지 읽음 처리
+export const postMarkBirthdayLetters = async () => {
+  const res = await api.post('/birthday/letters/me/mark');
+  return res.data;
+};
